dream_space_be/server/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
django-phonenumber-field==7.0.0
phonenumbers==8.13.0
Pillow==9.3.0